# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Rezi",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "operationId": "createUser",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Seu Nome"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "sua_senha"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "sua_senha"
                                    }
                                },
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ]
                            }
                        }
                    },
                    "responses": {
                        "201": {
                            "description": "Usuário criado com sucesso",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Usuário criado com sucesso"
                                            },
                                            "user": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Seu Nome"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "email@example.com"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "nullable": true,
                                                        "example": null
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-04T04:12:09.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-04T14:28:57.000000Z"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "name",
                                                    "email",
                                                    "created_at",
                                                    "updated_at"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "400": {
                            "description": "Erro ao criar usuário",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Erro ao criar usuário"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}